# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AKIS Platform API
  description: |
    Adaptive Knowledge Integrity System — AI Agent Workflows Engine.
    Scribe → Proto → Trace pipeline ile fikir → kod → test otomasyonu.
  version: 0.2.0
  contact:
    name: Ömer Yasir Önal
    url: https://akisflow.com
servers:
  - url: https://akisflow.com
    description: Production
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/settings/ai-keys:
    put:
      tags: [Settings]
      summary: Save/update AI API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider, apiKey]
              properties:
                provider: { type: string, enum: [anthropic, openai, openrouter] }
                apiKey: { type: string }
      responses:
        "200":
          description: Key saved (AES-256-GCM encrypted)
    delete:
      tags: [Settings]
      summary: Delete AI API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider]
              properties:
                provider: { type: string, enum: [anthropic, openai, openrouter] }
      responses:
        "200":
          description: Key deleted

  # ─── GITHUB ─────────────────────────────────────
...
  /api/github/repos:
    get:
      tags: [GitHub]
      summary: List user's GitHub repos
      responses:
        "200":
          description: Repo list
          content:
            application/json:
              schema:
                type: object
                properties:
                  repos:
                    type: array
                    items:
                      type: object
                      properties:
                        name: { type: string }
                        fullName: { type: string }
                        private: { type: boolean }
                        url: { type: string, format: uri }
        "403":
          description: GitHub not connected (not a session expiry)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code: { type: string, enum: [GITHUB_NOT_CONNECTED] }
                      message: { type: string }
  /api/github/connect:
    post:
      tags: [GitHub]
      summary: Connect GitHub via PAT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        "200":
          description: GitHub connected
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: akis_session
...
security:
  - cookieAuth: []